+++ app/Http/Controllers/InstructorController.php
@@ -22,6 +22,8 @@ class InstructorController extends Controller
      */
     public function create()
     {
+        $this->authorize('administrar');
+
         return view('instructors.create');
     }
 
@@ -30,6 +32,8 @@ class InstructorController extends Controller
      */
     public function store(StoreInstructorRequest $request)
     {
+        $this->authorize('administrar');
+
         Instructor::create($request->validated());
         return redirect()->route('instructors.index');
     }
@@ -47,6 +51,8 @@ class InstructorController extends Controller
      */
     public function edit(Instructor $instructor)
     {
+        $this->authorize('administrar');
+
         return view('instructors.edit', compact('instructor'));
     }
 
@@ -55,6 +61,8 @@ class InstructorController extends Controller
      */
     public function update(UpdateInstructorRequest $request, Instructor $instructor)
     {
+        $this->authorize('administrar');
+
         $instructor->update($request->validated());
         return redirect()->route('instructors.index');
     }
@@ -64,6 +72,8 @@ class InstructorController extends Controller
      */
     public function destroy(Instructor $instructor)
     {
+        $this->authorize('administrar');
+        
         $instructor->delete();
         return redirect()->route('instructors.index');
     }
